+++ src/Policy.ts
@@ -1,6 +1,5 @@
 import { IBreaker } from './breaker/Breaker';
 import { BulkheadPolicy } from './BulkheadPolicy';
-import { CancellationToken } from './CancellationToken';
 import { CircuitBreakerPolicy } from './CircuitBreakerPolicy';
 import { Event } from './common/Event';
 import { ExecuteWrapper } from './common/Executor';
@@ -60,10 +59,10 @@ export interface ISuccessEvent {
 
 export interface IDefaultPolicyContext {
   /**
-   * Cancellation token for the operation. This is propagated through multiple
+   * Abort signal for the operation. This is propagated through multiple
    * retry policies.
    */
-  cancellationToken: CancellationToken;
+  signal: AbortSignal;
 }
 
 /**
@@ -92,7 +91,7 @@ export interface IPolicy<
    */
   execute<T>(
     fn: (context: ContextType) => PromiseLike<T> | T,
-    cancellationToken?: CancellationToken,
+    signal?: AbortSignal,
   ): Promise<T | AltReturn>;
 }
 
@@ -200,15 +199,12 @@ export class Policy {
     return {
       onFailure: p[0].onFailure,
       onSuccess: p[0].onSuccess,
-      execute<T>(
-        fn: (context: C) => PromiseLike<T> | T,
-        cancellationToken: CancellationToken,
-      ): Promise<T | A> {
+      execute<T>(fn: (context: C) => PromiseLike<T> | T, signal: AbortSignal): Promise<T | A> {
         const run = (context: C, i: number): PromiseLike<T | A> | T | A =>
           i === p.length
             ? fn(context)
-            : p[i].execute(next => run({ ...context, ...next }, i + 1), context.cancellationToken);
-        return Promise.resolve(run({ cancellationToken } as C, 0));
+            : p[i].execute(next => run({ ...context, ...next }, i + 1), context.signal);
+        return Promise.resolve(run({ signal } as C, 0));
       },
     };
   }
@@ -303,8 +299,8 @@ export class Policy {
       }
 
       descriptor.value = function (this: unknown, ...args: any[]) {
-        const ct = args[args.length - 1] instanceof CancellationToken ? args.pop() : undefined;
-        return policy.execute(context => inner.apply(this, [...args, context]), ct);
+        const signal = args[args.length - 1] instanceof AbortSignal ? args.pop() : undefined;
+        return policy.execute(context => inner.apply(this, [...args, context]), signal);
       };
     };
   }